refactor: Extracted functions from release date
refs: BLAIS5-3143
