+++ cloud_functions/check_questionnaire_release_date.py
@@ -4,11 +4,6 @@ import logging
 from google.cloud import datastore, tasks_v2
 from datetime import datetime
 from typing import List, Coroutine, Any
-from google.protobuf.duration_pb2 import Duration
-from datetime import timedelta
-from uuid import uuid4
-
-from appconfig import Config
 from cloud_functions.logging import setup_logger
 from models.questionnaire_case_task_model import QuestionnaireCaseTaskModel
 
@@ -24,43 +19,6 @@ def create_tasks(
 async def run(task_requests: List[tasks_v2.CreateTaskRequest]) -> None:
     task_client = tasks_v2.CloudTasksAsyncClient()
     await asyncio.gather(*create_tasks(task_requests, task_client))
-
-
-def create_questionnaire_task_name(job_model: QuestionnaireCaseTaskModel) -> str:
-    return (
-        f"{job_model.questionnaire}-{str(uuid4())}"
-    )
-
-
-def prepare_questionnaire_tasks(
-        job_models: List[QuestionnaireCaseTaskModel],
-) -> List[tasks_v2.CreateTaskRequest]:
-
-    config = Config.from_env()
-
-    duration = Duration()
-    duration.FromTimedelta(timedelta(minutes=30))
-
-    task_requests = []
-    for job_model in job_models:
-        request = tasks_v2.CreateTaskRequest(
-            parent=config.totalmobile_jobs_queue_id,
-            task=tasks_v2.Task(
-                name=f"{config.totalmobile_jobs_queue_id}/tasks/{create_questionnaire_task_name(job_model)}",
-                http_request=tasks_v2.HttpRequest(
-                    http_method="POST",
-                    url=f"https://{config.region}-{config.gcloud_project}.cloudfunctions.net/{config.totalmobile_job_cloud_function}",
-                    body=job_model.json().encode(),
-                    headers={
-                        "Content-Type": "application/json",
-                    },
-                    oidc_token={"service_account_email": config.cloud_function_sa},
-                ),
-                dispatch_deadline=duration,
-            ),
-        )
-        task_requests.append(request)
-    return task_requests
 
 
 def map_questionnaire_case_task_models(questionnaires: List[str]) -> List[QuestionnaireCaseTaskModel]:
